home/backend/requirements.txt
uvicorn 
sqlalchemy 
jinja2
numpy
pandas
fastapi
langchain-together
dotenv
